llm-poc/data-harmonization/requirements.txt
requests
python-dotenv
openai